+++ R/add_flextable.R
@@ -21,8 +21,7 @@
 #' doc %>% add_text(title=title) %>%
 #'         add_flextable(ftable) %>%
 #'         add_text(title=title2) %>%
-#'         add_flextable(ft) %>%
-#'         print(target="mytable.docx")
+#'         add_flextable(ft)
 add_flextable=function(mydoc,ftable,echo=FALSE,code="",landscape=FALSE){
      if("mytable" %in% class(ftable)){
           ft<-mytable2flextable(ftable)