+++ R/data2office.R
@@ -162,17 +162,10 @@ data2office=function(data,
             mydoc=add_flextable(mydoc,ft,code=data$code[i],echo=echo1,landscape = landscape1)
         } else if(data$type[i]=="ggplot"){
             mydoc=add_ggplot(mydoc,code=data$code[i],preprocessing=preprocessing,top=ifelse(echo1,2,1.5))
-        }else if(data$type[i]=="2ggplots"){
-
-            codes=unlist(strsplit(data$code[i],"\n"))
-            # codes=unlist(strsplit(sampleData2$code[8],"\n"))
-            gg1=codes[1]
-            gg2=codes[2]
-            mydoc=add_2ggplots(mydoc,plot1=gg1,plot2=gg2,preprocessing=preprocessing,top=ifelse(echo1,2,1.5))
         } else if(data$type[i]=="plot"){
             mydoc<-add_plot(mydoc,data$code[i],preprocessing=preprocessing,top=ifelse(echo1,2,1.5))
 
-        } else if(data$type[i]=="2plots"){
+        } else if(data$type[i] %in% c("2plots","2ggplots")){
 
             codes=unlist(strsplit(data$code[i],"\n"))
             mydoc=add_2plots(mydoc,plotstring1=codes[1],plotstring2=codes[2],preprocessing=preprocessing,top=ifelse(echo1,2,1.5))
@@ -218,7 +211,7 @@ data2office=function(data,
 #' library(rrtable)
 #' library(moonBook)
 #' library(ggplot2)
-#' data2pptx(sampleData2)
+#' data2pptx(sampleData2,echo=TRUE)
 #' }
 data2pptx=function(...){
     data2office(...)
@@ -232,7 +225,7 @@ data2pptx=function(...){
 #' library(rrtable)
 #' library(moonBook)
 #' library(ggplot2)
-#' data2docx(sampleData2)
+#' data2docx(sampleData2,echo=TRUE)
 #' }
 data2docx=function(...){
     data2office(...,format="docx")